% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvoFreq.R
\name{long_to_wide_size_df}
\alias{long_to_wide_size_df}
\title{Convert XYZ}
\usage{
long_to_wide_size_df(edges_df, long_pop_sizes_df, time_col_name, clone_col_name,
  parent_col_name, size_col_name)
}
\arguments{
\item{edges_df}{Description here.}

\item{long_pop_sizes_df}{Description here.}

\item{time_col_name}{File to plot the animation to.}

\item{clone_col_name}{Description here.}

\item{parent_col_name}{Description here.}

\item{size_col_name}{Description here.}
}
\description{
Builds...
}
\examples{
## Using Provided Example Data
## Data is in a long format consisting of two files:
## 1. Population Data Frame
head(pop_df)
## 2. Edges Data Frame
head(edges_df)

## For EvoFreq to use this data it must convert it.
wide_df <- long_to_wide_size_df(edges_df, pop_df, "Generation", "Identity", "Parent", "Population")

## Pull apart the necessary information for creating the frequency dynamics object
clone_list <- wide_df$clones
parent_list <- wide_df$parents
clone_df <- wide_df$wide_size_df
timepoints = sort(unique(pop_df$Generation))

## Get the frequency dynamics object using get_freq_dynamics()

## Plot the frequency dynamics using plot_freq_dynamics()
}
\keyword{long_to_wide_size_df}
